From django-dropbox to django-imgur
though is not a real storage solution, imgur can be used as a convenient
solution to drop images. Every album is managed as a path, and every
file got as name is "absolute" path.

Since my use now is as a drop-and-forget storage, the only tested
feature is the upload/creation of files and "directories". More to come
next